LexiStream-feedback/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
python-dotenv==1.0.0
pydub==0.25.1
Werkzeug==2.3.7
SpeechRecognition==3.10.0
PyMySQL==1.1.0
cohere>=5.0.0